requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
python-dotenv==1.0.0
pymongo==4.5.0
motor==3.3.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
openai==1.3.5
pillow==10.1.0
pandas==2.1.3
numpy==1.26.2
pydantic==2.5.2
pydantic-settings==2.1.0
jinja2==3.1.2
